setup_folder/requirements.txt
kafka-python
snowflake-connector-python
pandas